chore: cleanup multicodec port
