+++ src/multicodec/multicodec.c
@@ -1,4 +1,4 @@
-#include "multicodec.h"
+#include "multicodec/multicodec.h"
 
 int multicodec_encode(const char *codecName, char *inData, size_t inDataLength,
                       char *outData, size_t maxOutDataLength) {
